refactor: Move, document, and rename function. Also adds explicit `ReleaseLevel` type.
