+++ tools/api-markdown-documenter/src/index.ts
@@ -82,6 +82,7 @@ export type {
 	ApiModifier,
 	ApiModuleLike,
 	ApiSignatureLike,
+	ReleaseLevel,
 	ValidApiItemKind,
 } from "./utilities/index.js";
